Reimplement `precise-hologram-movement` differently to fix lag issues
The `precise-hologram-movement` config option was creating lag on some servers. The option is now removed from the configuration and its functionality has been reimplemented through ProtocolLib (when installed). It now uses a different method that doesn't rely on inspecting stacktraces.

Secondly, a workaround was added to fix the Minecraft bug causing entities to not move client side when teleported shortly after spawning. The cause is EntityTrackerEntry (a NMS class) not sending the teleport packet.